LPS-144170 Move org.monte usages to imageio-plugins
