+++ modules/apps/imageio-plugins/imageio-plugins/build.gradle
@@ -1,4 +1,5 @@
 dependencies {
+	compileInclude group: "com.liferay", name: "org.monte", version: "0.7.7"
 	compileInclude group: "com.twelvemonkeys.common", name: "common-image", version: "3.3.2"
 	compileInclude group: "com.twelvemonkeys.common", name: "common-io", version: "3.3.2"
 	compileInclude group: "com.twelvemonkeys.common", name: "common-lang", version: "3.3.2"
@@ -18,6 +19,7 @@ dependencies {
 	compileInclude group: "com.twelvemonkeys.imageio", name: "imageio-thumbsdb", version: "3.5"
 	compileInclude group: "com.twelvemonkeys.imageio", name: "imageio-tiff", version: "3.5"
 
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
 	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
 }
