+++ modules/apps/imageio-plugins/imageio-plugins/src/main/java/com/liferay/imageio/plugins/internal/activator/ImageIOPluginsBundleActivator.java
@@ -14,6 +14,8 @@
 
 package com.liferay.imageio.plugins.internal.activator;
 
+import com.liferay.portal.kernel.util.ArrayUtil;
+
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.Set;
@@ -22,6 +24,8 @@ import javax.imageio.spi.IIORegistry;
 import javax.imageio.spi.ImageReaderSpi;
 import javax.imageio.spi.ImageWriterSpi;
 import javax.imageio.spi.ServiceRegistry;
+
+import org.monte.media.jpeg.CMYKJPEGImageReaderSpi;
 
 import org.osgi.framework.BundleActivator;
 import org.osgi.framework.BundleContext;
@@ -35,12 +39,47 @@ public class ImageIOPluginsBundleActivator implements BundleActivator {
 	public void start(BundleContext bundleContext) {
 		_register(ImageReaderSpi.class, _imageReaderSpiSet);
 		_register(ImageWriterSpi.class, _imageWriterSpiSet);
+
+		orderImageReaderSpis();
 	}
 
 	@Override
 	public void stop(BundleContext bundleContext) {
 		_unregister(_imageReaderSpiSet);
 		_unregister(_imageWriterSpiSet);
+	}
+
+	protected void orderImageReaderSpis() {
+		IIORegistry defaultIIORegistry = IIORegistry.getDefaultInstance();
+
+		ImageReaderSpi firstImageReaderSpi = null;
+		ImageReaderSpi secondImageReaderSpi = null;
+
+		Iterator<ImageReaderSpi> iterator =
+			defaultIIORegistry.getServiceProviders(ImageReaderSpi.class, true);
+
+		while (iterator.hasNext()) {
+			ImageReaderSpi imageReaderSpi = iterator.next();
+
+			if (imageReaderSpi instanceof CMYKJPEGImageReaderSpi) {
+				secondImageReaderSpi = imageReaderSpi;
+			}
+			else {
+				String[] formatNames = imageReaderSpi.getFormatNames();
+
+				if (ArrayUtil.contains(formatNames, "jpg", true) ||
+					ArrayUtil.contains(formatNames, "jpeg", true)) {
+
+					firstImageReaderSpi = imageReaderSpi;
+				}
+			}
+		}
+
+		if ((firstImageReaderSpi != null) && (secondImageReaderSpi != null)) {
+			defaultIIORegistry.setOrdering(
+				ImageReaderSpi.class, firstImageReaderSpi,
+				secondImageReaderSpi);
+		}
 	}
 
 	private <T> void _register(Class<T> clazz, Set<T> registeredProviders) {
