+++ modules/apps/imageio-plugins/imageio-plugins/src/main/resources/META-INF/services/javax.imageio.spi.ImageReaderSpi
